# Python benchmark for transform logging over time